requirements.txt
flask
pandas
openai
faiss-cpu
python-docx
pymupdf
tiktoken
rapidfuzz
